refactor: migrate to ESM and update project configuration
- Replace ts-node with tsx for ESM compatibility
- Update jest and typescript configurations for ESM support
- Simplify redis service initialization
- Consolidate type imports under @/types
- Update package.json scripts and dependencies
- Remove husky pre-commit hooks
- Add esbuild to workspace dependencies
