--- .husky/pre-commit
@@ -1,2 +1,0 @@
-pnpm lint-staged
-pnpm test
