+++ eslint.config.mjs
@@ -25,6 +25,7 @@ export default [
       "eslint.config.mjs",
       "coverage**/*",
       "*.history",
+      "jest.config.mjs",
     ],
   },
   ...compat.extends("eslint:recommended", "plugin:@typescript-eslint/recommended"),
