+++ jest.config.ts
@@ -1,18 +1,23 @@
-import { type Config } from "jest";
+import type { Config } from "jest";
+import type { JestConfigWithTsJest } from "ts-jest";
+import { ESM_TS_TRANSFORM_PATTERN, TS_EXT_TO_TREAT_AS_ESM } from "ts-jest";
 
-const config: Config = {
+const config: JestConfigWithTsJest = {
   verbose: true,
-  preset: "ts-jest",
-  moduleFileExtensions: ["js", "ts", "json", "node"],
   testTimeout: 10000,
-  modulePathIgnorePatterns: ["<rootDir>/dist"],
-  transformIgnorePatterns: ["node_modules/(?!(jest-)?ts-jest)"],
-  roots: ["<rootDir>/src"],
-  setupFilesAfterEnv: ["<rootDir>/src/jest.setup.ts"],
-  testEnvironment: "node",
-  moduleNameMapper: {
-    "^@/(.*)$": "<rootDir>/src/$1",
+  transform: {
+    [ESM_TS_TRANSFORM_PATTERN]: [
+      "ts-jest",
+      {
+        useESM: true,
+      },
+    ],
   },
-};
+  extensionsToTreatAsEsm: [...TS_EXT_TO_TREAT_AS_ESM],
+  moduleNameMapper: {
+    "^(\\.{1,2}/.*)\\.js$": "$1",
+  },
+  setupFilesAfterEnv: ["<rootDir>/src/jest.setup.ts"], //TODO fix this
+} satisfies Config;
 
 export default config;
