+++ nodemon.json
@@ -1,10 +1,7 @@
 {
   "watch": ["src"],
   "ext": ".ts",
-  "ignore": [],
-  "exec": "ts-node -r tsconfig-paths/register -r dotenv/config ./src/index.ts",
-  "env": {
-    "NODE_ENV": "development"
-  },
+  "ignore": ["dist"],
+  "exec": "tsx watch src/index.ts",
   "verbose": true
 }
