+++ package.json
@@ -9,12 +9,12 @@
     "local": "cross-env NODE_ENV=local nodemon",
     "local:dev": "cross-env NODE_ENV=development nodemon",
     "local:prod": "cross-env NODE_ENV=production pnpm prod",
-    "generate-keys": "ts-node scripts/generate-keys.ts",
-    "generate-env": "ts-node scripts/generate-env-files.ts",
-    "migrate": "ts-node -r tsconfig-paths/register src/config/migrate.ts",
+    "generate-keys": "tsx scripts/generate-keys.ts",
+    "generate-env": "tsx scripts/generate-env-files.ts",
+    "migrate": "tsx -r tsconfig-paths/register src/config/migrate.ts",
     "migrate:local": "cross-env NODE_ENV=local pnpm migrate",
     "migrate:prod": "cross-env NODE_ENV=production pnpm migrate",
-    "setup": "ts-node scripts/setup-project.ts",
+    "setup": "tsx scripts/setup-project.ts",
     "lint:fix": "eslint --fix .",
     "lint": "eslint src .",
     "prepare": "husky",
@@ -24,7 +24,7 @@
     "prod": "pnpm build && pnpm start",
     "start": "node dist/index.js",
     "test:coverage": "jest --runInBand --detectOpenHandles --forceExit --coverage",
-    "test": "cross-env NODE_ENV=test jest --runInBand --detectOpenHandles --forceExit"
+    "test": "cross-env NODE_ENV=test node --experimental-vm-modules node_modules/jest/bin/jest.js --runInBand --detectOpenHandles --forceExit"
   },
   "keywords": [
     "user",
@@ -37,18 +37,18 @@
     "bcrypt": "6.0.0",
     "cors": "2.8.5",
     "date-fns": "4.1.0",
-    "dotenv": "17.2.1",
+    "dotenv": "17.2.2",
     "express": "5.1.0",
-    "express-rate-limit": "8.0.1",
+    "express-rate-limit": "8.1.0",
     "helmet": "8.1.0",
     "ioredis": "5.7.0",
     "jsonwebtoken": "9.0.2",
     "maxmind": "^5.0.0",
-    "mongoose": "8.17.1",
+    "mongoose": "8.18.0",
     "morgan": "1.10.1",
     "passport": "0.7.0",
     "passport-google-oauth20": "2.0.0",
-    "pino": "9.8.0",
+    "pino": "9.9.2",
     "pino-http": "10.5.0",
     "socket.io": "4.8.1",
     "ua-parser-js": "^2.0.4",
@@ -56,7 +56,7 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "3.3.1",
-    "@eslint/js": "9.32.0",
+    "@eslint/js": "9.34.0",
     "@trivago/prettier-plugin-sort-imports": "5.2.2",
     "@types/bcrypt": "6.0.0",
     "@types/cors": "2.8.19",
@@ -67,25 +67,27 @@
     "@types/passport": "1.0.17",
     "@types/passport-google-oauth20": "2.0.16",
     "@types/supertest": "6.0.3",
-    "@typescript-eslint/eslint-plugin": "8.39.0",
-    "@typescript-eslint/parser": "8.39.0",
+    "@typescript-eslint/eslint-plugin": "8.42.0",
+    "@typescript-eslint/parser": "8.42.0",
     "cross-env": "10.0.0",
-    "eslint": "9.32.0",
+    "eslint": "9.34.0",
     "eslint-plugin-check-file": "3.3.0",
     "globals": "16.3.0",
     "husky": "9.1.7",
-    "jest": "30.0.5",
-    "lint-staged": "16.1.4",
+    "jest": "30.1.3",
+    "jest-environment-jsdom": "^30.1.2",
+    "lint-staged": "16.1.6",
     "nodemon": "3.1.10",
     "pino-pretty": "13.1.1",
     "prettier": "3.6.2",
     "supertest": "7.1.4",
     "ts-jest": "29.4.1",
-    "ts-node": "10.9.2",
+    "ts-node": "^10.9.2",
     "tsc-alias": "^1.8.16",
     "tsconfig-paths": "4.2.0",
+    "tsx": "^4.20.5",
     "typescript": "5.9.2",
-    "typescript-eslint": "8.39.0"
+    "typescript-eslint": "8.42.0"
   },
   "husky": {
     "hooks": {
