+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 onlyBuiltDependencies:
   - bcrypt
+  - esbuild
   - unrs-resolver
