+++ src/api/v1/user/user.model.ts
@@ -1,10 +1,12 @@
-import { Schema, model } from "mongoose";
+import mongoose, { Document, Schema, model } from "mongoose";
 
 import { hashPassword } from "@/utils";
 
 import type { User } from "./user.validation.js";
 
-export interface IUser extends Document, User {}
+export interface IUser extends Document, User {
+  id: mongoose.Types.ObjectId  
+}
 
 const userSchema = new Schema<IUser>(
   {
