+++ src/utils/test/auth.utils.ts
@@ -2,11 +2,10 @@ import supertest, { type Response } from "supertest";
 
 import { success } from "@/api/v1/auth/auth.constant.js";
 import type { Login, SignUpResponse } from "@/api/v1/auth/auth.validation.js";
-import { app } from "@/app.js";
-import { ErrorTypeEnum, STATUS_CODES, errorMap } from "@/constants";
-import { NotificationType } from "@/constants/events.constant.js";
+import { ErrorTypeEnum, NotificationType, STATUS_CODES, errorMap } from "@/constants";
 
 import { type ResetPassword } from "../../api/v1/user/user.validation.js";
+import { app } from "../../app.js";
 import { expectOTPVerificationSuccess, retrieveOTP, verifyOTP } from "./otp.utils.js";
 
 export const login = async (loginData: Omit<Login, "deviceInfo">): Promise<Response> => {
