+++ src/utils/test/otp.utils.ts
@@ -3,9 +3,10 @@ import supertest, { type Response } from "supertest";
 import { success } from "@/api/v1/auth/auth.constant.js";
 import { OtpService } from "@/api/v1/otp/otp.service.js";
 import type { OtpEvent, OtpSchema, OtpVerifyEvent } from "@/api/v1/otp/otp.validation.js";
-import { app } from "@/app.js";
 import { NotificationType, STATUS_CODES } from "@/constants";
 import { type MongooseObjectId } from "@/types";
+
+import { app } from "../../app.js";
 
 export const requestOTP = async (otpEvents: OtpEvent): Promise<Response> => {
   return supertest(app).post("/api/v1/otp").send(otpEvents);
