+++ src/utils/test/permission.utils.ts
@@ -2,8 +2,9 @@ import supertest, { type Response } from "supertest";
 
 import { success } from "@/api/v1/permission/permission.constant.js";
 import type { CreatePermission } from "@/api/v1/permission/permission.validation.js";
-import { app } from "@/app.js";
 import { STATUS_CODES } from "@/constants";
+
+import { app } from "../../app.js";
 
 export const createPermissionRequest = async (permission: CreatePermission): Promise<Response> => {
   return await supertest(app).post("/api/v1/permissions").send(permission);
