+++ src/utils/test/role-permission.utils.ts
@@ -2,8 +2,9 @@ import supertest, { type Response } from "supertest";
 
 import { success } from "@/api/v1/rolePermission/role-permission.constant.js";
 import type { RolePermission } from "@/api/v1/rolePermission/role-permission.validation.js";
-import { app } from "@/app.js";
 import { STATUS_CODES } from "@/constants";
+
+import { app } from "../../app.js";
 
 export const createRolePermissionRequest = async (rolePermission: RolePermission, authorizationHeader: string) => {
   return await supertest(app)
